Add support for labels and taints to HostedCP NodePools
This covers both creation and modification. The UX should be
the same as Classic ROSA.

Related: SDA-8219
