+++ cmd/create/machinepool/helper.go
@@ -79,3 +79,53 @@ func getSubnetOptions(r *rosa.Runtime, cluster *cmv1.Cluster) ([]string, error) 
 
 	return subnetOptions, nil
 }
+
+func getTaints(cmd *cobra.Command, r *rosa.Runtime) []*cmv1.TaintBuilder {
+	taints := args.taints
+	if interactive.Enabled() {
+		var err error
+		taints, err = interactive.GetString(interactive.Input{
+			Question: "Taints",
+			Help:     cmd.Flags().Lookup("taints").Usage,
+			Default:  taints,
+			Validators: []interactive.Validator{
+				taintValidator,
+			},
+		})
+		if err != nil {
+			r.Reporter.Errorf("Expected a valid comma-separated list of attributes: %s", err)
+			os.Exit(1)
+		}
+	}
+	taintBuilders, err := parseTaints(taints)
+	if err != nil {
+		r.Reporter.Errorf("%s", err)
+		os.Exit(1)
+	}
+	return taintBuilders
+}
+
+func getLabelMap(cmd *cobra.Command, r *rosa.Runtime) map[string]string {
+	labels := args.labels
+	if interactive.Enabled() {
+		var err error
+		labels, err = interactive.GetString(interactive.Input{
+			Question: "Labels",
+			Help:     cmd.Flags().Lookup("labels").Usage,
+			Default:  labels,
+			Validators: []interactive.Validator{
+				LabelValidator,
+			},
+		})
+		if err != nil {
+			r.Reporter.Errorf("Expected a valid comma-separated list of attributes: %s", err)
+			os.Exit(1)
+		}
+	}
+	labelMap, err := ParseLabels(labels)
+	if err != nil {
+		r.Reporter.Errorf("%s", err)
+		os.Exit(1)
+	}
+	return labelMap
+}
