+++ cmd/edit/machinepool/machinepool.go
@@ -1,10 +1,8 @@
 package machinepool
 
 import (
-	"fmt"
 	"os"
 	"regexp"
-	"strings"
 
 	cmv1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
 	"github.com/openshift/rosa/pkg/interactive"
@@ -136,38 +134,7 @@ func editMachinePool(cmd *cobra.Command, machinePoolID string, clusterKey string
 
 	labelMap := getLabels(cmd, r.Reporter, machinePool.Labels())
 
-	taints := args.taints
-	taintBuilders := []*cmv1.TaintBuilder{}
-	if interactive.Enabled() {
-		if taints == "" {
-			for _, taint := range machinePool.Taints() {
-				if taints != "" {
-					taints += ","
-				}
-				taints += fmt.Sprintf("%s=%s:%s", taint.Key(), taint.Value(), taint.Effect())
-			}
-		}
-		taints, err = interactive.GetString(interactive.Input{
-			Question: "Taints",
-			Help:     cmd.Flags().Lookup("taints").Usage,
-			Default:  taints,
-		})
-		if err != nil {
-			r.Reporter.Errorf("Expected a valid comma-separated list of attributes: %s", err)
-			os.Exit(1)
-		}
-	}
-	taints = strings.Trim(taints, " ")
-	if taints != "" {
-		for _, taint := range strings.Split(taints, ",") {
-			if !strings.Contains(taint, "=") || !strings.Contains(taint, ":") {
-				r.Reporter.Errorf("Expected key=value:scheduleType format for taints")
-				os.Exit(1)
-			}
-			tokens := strings.FieldsFunc(taint, Split)
-			taintBuilders = append(taintBuilders, cmv1.NewTaint().Key(tokens[0]).Value(tokens[1]).Effect(tokens[2]))
-		}
-	}
+	taintBuilders := getTaints(cmd, r, machinePool.Taints())
 
 	mpBuilder := cmv1.NewMachinePool().
 		ID(machinePool.ID())
@@ -324,44 +291,4 @@ func Split(r rune) bool {
 // Single-AZ: AvailabilityZones == []string{"us-east-1a"}
 func isMultiAZMachinePool(machinePool *cmv1.MachinePool) bool {
 	return len(machinePool.AvailabilityZones()) != 1
-}
-
-func getLabels(cmd *cobra.Command,
-	reporter *rprtr.Object,
-	existingLabels map[string]string) map[string]string {
-	var err error
-	labels := args.labels
-	labelMap := make(map[string]string)
-	if interactive.Enabled() {
-		if labels == "" {
-			for lk, lv := range existingLabels {
-				if labels != "" {
-					labels += ","
-				}
-				labels += fmt.Sprintf("%s=%s", lk, lv)
-			}
-		}
-		labels, err = interactive.GetString(interactive.Input{
-			Question: "Labels",
-			Help:     cmd.Flags().Lookup("labels").Usage,
-			Default:  labels,
-		})
-		if err != nil {
-			reporter.Errorf("Expected a valid comma-separated list of attributes: %s", err)
-			os.Exit(1)
-		}
-	}
-
-	labels = strings.Trim(labels, " ")
-	if labels != "" {
-		for _, label := range strings.Split(labels, ",") {
-			if !strings.Contains(label, "=") {
-				reporter.Errorf("Expected key=value format for labels")
-				os.Exit(1)
-			}
-			tokens := strings.Split(label, "=")
-			labelMap[strings.TrimSpace(tokens[0])] = strings.TrimSpace(tokens[1])
-		}
-	}
-	return labelMap
 }
